Rename entities class package
