+++ src/main/java/ch/heigvd/amt/backend/entities/Category.java
@@ -1,4 +1,4 @@
-package ch.heigvd.amt.backend.DBSchema;
+package ch.heigvd.amt.backend.entities;
 
 import lombok.Data;
