+++ src/main/java/ch/heigvd/amt/backend/repository/CategoryDAO.java
@@ -1,7 +1,6 @@
 package ch.heigvd.amt.backend.repository;
 
-import ch.heigvd.amt.backend.DBSchema.Category;
-import ch.heigvd.amt.backend.DBSchema.Product;
+import ch.heigvd.amt.backend.entities.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryDAO extends JpaRepository<Category, Integer> {
