+++ src/main/java/ch/heigvd/amt/backend/repository/ProductDAO.java
@@ -1,6 +1,6 @@
 package ch.heigvd.amt.backend.repository;
 
-import ch.heigvd.amt.backend.DBSchema.Product;
+import ch.heigvd.amt.backend.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
